pyproject.toml
[tool.poetry]
name = "langchain-task-template"
version = "0.1.0"
description = ""
authors = ["TimeSurge Labs <chandler@timesurgelabs.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.7"
langchain-openai = "^0.0.6"
python-dotenv = "^1.0.1"
fire = "^0.5.0"
pydantic = "^2.6.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
